+++ Kong/src/entities/Sparks.cpp
@@ -50,7 +50,7 @@ void Sparks::launchSpark() {
 
 }
 
-void Sparks::setDelayMax(uint8_t delayMax, bool updateDelay) {
+void Sparks::setDelayMax(uint16_t delayMax, bool updateDelay) {
 
     if (updateDelay) this->delay = random(0, delayMax / 2);
     this->delayMax = delayMax;
